[TASK] Use default pageTitle with custom wrap
The title tag is no longer built in 'page.headerData'.
Instead it relies on 'config.pageTitleProviders'.

Until it's possible to override the sitetitle in the
TypoScript template (sys_template) through TypoScript,
a TypoScript constant is used to append the website name.

This appendix will not be added if a page or news article
has a SEO title or an alternative title set.