[FEATURE] Rework BackendLayouts
This starter kit now provides five BackendLayouts with
corresponding Fluid templates.

Templates with multiple columns have additional full-width
columns on the top and bottom to be more flexible.

The example stylesheet contains some declarations from the
inuitcss framework just to provide a working columns layout
in the frontend. Of course you'll have to replace this with
your own stylesheet (and templates).