[TASK] Update title tag generation
The website title from Site Configuration will be appended to the
title tag. The custom configuration allows to omit the website title,
if a page (or news records) has an optimized title for search engines.

Page title separator is set directly in noTrimWrap.

The former custom TypoScript constant for the website name is removed.